📦 JE-62-chore: 파일경로 변경 및 삭제기능 추가
