+++ src/pages/mypage/myActivities/index.tsx
@@ -1,11 +1,12 @@
-import MyActivityForm from '@/components/ActivitiyForm';
-import MyCard from '@/components/activitiesManagement/MyCard';
+import MyCard from '@/components/activitiesManagement/myCard';
 import Button from '@/components/commons/Button';
 import Modal from '@/components/commons/Modal';
+import MyPageLayout from '@/components/mypage/MyPageLayout';
 import { getMyActivities } from '@/libs/api/myActivities';
 import { useInfiniteQuery } from '@tanstack/react-query';
 import Image from 'next/image';
-import React, { useEffect, useMemo, useRef, useState } from 'react';
+import { useRouter } from 'next/navigation';
+import React, { useEffect, useMemo, useRef } from 'react';
 
 import EmptyImage from '/public/imgs/empty.png';
 
@@ -20,8 +21,8 @@ const useMyActivities = (size = 20) => {
 };
 
 export default function MyActivities() {
-  const [showActivityForm, setShowActivityForm] = useState(false);
   const lastCardRef = useRef<HTMLDivElement | null>(null);
+  const router = useRouter();
 
   const { data, fetchNextPage, hasNextPage, isFetchingNextPage, status } =
     useMyActivities();
@@ -64,55 +65,55 @@ export default function MyActivities() {
   }, [fetchNextPage, hasNextPage, isFetchingNextPage]);
 
   const handleConfirm = () => {
-    setShowActivityForm(true);
+    router.push('myActivities/registerForm');
   };
 
   if (status === 'pending') return <div>로딩 중...</div>;
   if (status === 'error') return <div>에러가 발생했습니다.</div>;
 
-  if (showActivityForm) {
-    return <MyActivityForm />;
-  }
-
   return (
-    <div className="mb-50">
-      <div className="mb-24 flex items-center justify-between">
-        <h1 className="text-3xl-bold">내 체험 관리</h1>
-        <Modal.Root>
-          <Modal.Trigger>
-            <Button className="max-w-120 rounded-md px-16 py-10">
-              체험 등록하기
-            </Button>
-          </Modal.Trigger>
-          <Modal.Content>
-            <Modal.Description>체험을 등록하시겠습니까?</Modal.Description>
-            <Modal.Close onConfirm={handleConfirm} confirm>
-              확인
-            </Modal.Close>
-          </Modal.Content>
-        </Modal.Root>
-      </div>
-
-      {sortedActivities.length === 0 ? (
-        <div className="flex flex-col items-center justify-center">
-          <Image src={EmptyImage} alt="빈 이미지" width={200} height={200} />
-          <p className="mt-20 text-2xl-medium text-custom-gray-700">
-            아직 등록한 체험이 없어요
-          </p>
+    <MyPageLayout>
+      <div className="mb-50">
+        <div className="mb-24 flex items-center justify-between">
+          <h1 className="text-3xl-bold">내 체험 관리</h1>
+          <Modal.Root>
+            <Modal.Trigger>
+              <Button className="max-w-120 rounded-md px-16 py-10">
+                체험 등록하기
+              </Button>
+            </Modal.Trigger>
+            <Modal.Content>
+              <Modal.Description className="text-center">
+                체험을 등록하시겠습니까?
+              </Modal.Description>
+              <Modal.Close onConfirm={handleConfirm} confirm>
+                예
+              </Modal.Close>
+            </Modal.Content>
+          </Modal.Root>
         </div>
-      ) : (
-        <>
-          {sortedActivities.map((activity, index) => (
-            <div
-              ref={index === sortedActivities.length - 1 ? lastCardRef : null}
-              key={activity.id}
-            >
-              <MyCard {...activity} />
-            </div>
-          ))}
-          {isFetchingNextPage && <div>로딩 중...</div>}
-        </>
-      )}
-    </div>
+
+        {sortedActivities.length === 0 ? (
+          <div className="flex flex-col items-center justify-center">
+            <Image src={EmptyImage} alt="빈 이미지" width={200} height={200} />
+            <p className="mt-20 text-2xl-medium text-custom-gray-700">
+              아직 등록한 체험이 없어요
+            </p>
+          </div>
+        ) : (
+          <>
+            {sortedActivities.map((activity, index) => (
+              <div
+                ref={index === sortedActivities.length - 1 ? lastCardRef : null}
+                key={activity.id}
+              >
+                <MyCard {...activity} />
+              </div>
+            ))}
+            {isFetchingNextPage && <div>로딩 중...</div>}
+          </>
+        )}
+      </div>
+    </MyPageLayout>
   );
 }
